Remove compiler warnings on Solaris 8 (Python's 2.59).

+++ extensions/pyexpat.c
@@ -1615,7 +1615,7 @@ get_version_string(void)
     char *rev = rcsid;
     int i = 0;
 
-    while (!isdigit(*rev))
+    while (!isdigit((int)*rev))
         ++rev;
     while (rev[i] != ' ' && rev[i] != '\0')
         ++i;
